requirements.txt
# requirements.txt
streamlit==1.26.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==3.3.5
shap==0.42.1
lime==0.2.0.1
reportlab==4.0.4
matplotlib==3.7.2
plotly==5.16.1
dash==2.13.0
google-cloud-translate==3.11.3
pdfkit==1.0.0
openpyxl==3.1.2
pillow==10.0.0